Seção 17: Palestrantes e Redes Sociais - Frontend

+++ Back/src/ProEventos.API/Helpers/Util.cs
@@ -30,8 +30,10 @@ namespace ProEventos.API.Helpers
             return imageName;
         }
         public void DeleteImage(string imageName, string destino){
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"resources/{destino}", imageName);
-            if(System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
+            if(!string.IsNullOrEmpty(imageName)){
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"resources/{destino}", imageName);
+                if(System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
+            }
         }
     }
 }
